remove explicit allowBackup
The following line from the android manifest forces users of this library to utilize tools:replace="android:allowBackup" when switiching to android:allowBackup="false" in their manifest:

     <application ... android:allowBackup="true" ...>

`allowBackup=true` is implicit. If there's no special reason for having this explicitly added to the lib's manifest I'd suggest to drop it.