Fixed some obvious errors in the translation
Some translations are wrong. i.e. "Turkey" is translated into the literal meaning of "The turkey animal" in Chinese. And couple of other names which are not correct are updated.

PS: same problem would occur if the traditional Chinese xml file is a direct translation from the this chinese_simplified one. Unfortunately I didn't have time to check that.